Extract player events
